Use separated styles for each component/container
